+++ app/shortcuts.cpp
@@ -1,6 +1,6 @@
 #include "shortcuts.h"
 
-Shortcuts::Shortcuts(QWidget *parent, Config::Global &global_c) : QObject(parent), global_conf(global_c),
+Shortcuts::Shortcuts(QWidget *parent) : QObject(parent),
   _quit(parent),
   _focus_library(parent),
   _focus_playlists(parent),
@@ -29,8 +29,11 @@ Shortcuts::Shortcuts(QWidget *parent, Config::Global &global_c) : QObject(parent
 
 QVector<QPair<QString, QString> > Shortcuts::describe() const {
   QVector<QPair<QString, QString> > r;
-  r << QPair<QString, QString>(tr("Open shortcuts dialog"), _open_shortcuts_menu.key().toString());
-  r << QPair<QString, QString>(tr("Quit"), _quit.key().toString());
+  r << QPair<QString, QString>(tr("Play"), _play.key().toString());
+  r << QPair<QString, QString>(tr("Stop"), _stop.key().toString());
+  r << QPair<QString, QString>(tr("Pause"), _pause.key().toString());
+  r << QPair<QString, QString>(tr("Next"), _next.key().toString());
+  r << QPair<QString, QString>(tr("Previous"), _prev.key().toString());
   r << QPair<QString, QString>(tr("Focus on library"), _focus_library.key().toString());
   r << QPair<QString, QString>(tr("Focus on playlists"), _focus_playlists.key().toString());
   r << QPair<QString, QString>(tr("Focus on playlist"), _focus_playlist.key().toString());
@@ -40,11 +43,8 @@ QVector<QPair<QString, QString> > Shortcuts::describe() const {
   r << QPair<QString, QString>(tr("Open main menu"), _open_main_menu.key().toString());
   r << QPair<QString, QString>(tr("Open playback log"), _open_playback_log.key().toString());
   r << QPair<QString, QString>(tr("Open sort menu"), _open_sort_menu.key().toString());
-  r << QPair<QString, QString>(tr("Play"), _play.key().toString());
-  r << QPair<QString, QString>(tr("Stop"), _stop.key().toString());
-  r << QPair<QString, QString>(tr("Pause"), _pause.key().toString());
-  r << QPair<QString, QString>(tr("Next"), _next.key().toString());
-  r << QPair<QString, QString>(tr("Previous"), _prev.key().toString());
+  r << QPair<QString, QString>(tr("Open shortcuts dialog"), _open_shortcuts_menu.key().toString());
+  r << QPair<QString, QString>(tr("Quit"), _quit.key().toString());
   return r;
 }