+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [1.0.15] - NEXT
 ### Added
 - "Save settings" main menu item to force saving settings and playlists to both config files
+- Ability to select playlist via enter key
+- Added "Search on web" track dialog menu item
 
 ## [1.0.14] - 2021-08-16
 ### Added